implementing hide button functionality
Using ternary operator for adding hideButton functionality.The `hideButtons` prop is used to conditionally hide or show the buttons in the `Search` component. When `hideButtons` is set to `true`, the buttons are hidden, and when it's set to `false` (or not provided), the buttons are shown. This allows for dynamic control over the visibility of the buttons based on the specific requirements or conditions of the application.

using useNavigate() hook to navigate to the search page